+++ packages/synapse-react-client/src/components/DataGrid/SynapseGrid.tsx
@@ -43,7 +43,7 @@ import {
 } from './utils/DataGridUtils'
 import { StartGridSession } from './StartGridSession'
 import { useDataGridWebSocket } from './useDataGridWebsocket'
-import { useGridUndo } from './hooks/useGridUndo'
+import { useGridUndoRedo } from './hooks/useGridUndoRedo'
 import { applyModelChange, ModelChange } from './utils/applyModelChange'
 import { mapOperationsToModelChanges } from './utils/mapOperationsToModelChanges'
 
@@ -190,6 +190,9 @@ const SynapseGrid = forwardRef<
     operations = removeNoOpOperations(newValue, rowValues, operations)
 
     if (operations.length > 0) {
+      // Clear redo stack since new changes invalidate redo history
+      clearRedoStack()
+
       // Track row creation, updates, and deletions to keep UI state and undo history in sync
 
       // Add all operations to the undo stack
@@ -206,7 +209,7 @@ const SynapseGrid = forwardRef<
     }
   }
 
-  const applyModelChangeFromUndo = useCallback(
+  const applyModelChangeFromUndoRedo = useCallback(
     (change: ModelChange) => {
       if (!model) {
         console.error('Model is not initialized')
@@ -223,9 +226,8 @@ const SynapseGrid = forwardRef<
     [model, applyAndCommitChanges],
   )
 
-  const { undoUI, addOperationsToUndoStack } = useGridUndo(
-    applyModelChangeFromUndo,
-  )
+  const { undoUI, redoUI, addOperationsToUndoStack, clearRedoStack } =
+    useGridUndoRedo(applyModelChangeFromUndoRedo)
 
   // Track the currently selected row index
   const [selectedRowIndex, setSelectedRowIndex] = useState<number | null>(null)
@@ -300,6 +302,7 @@ const SynapseGrid = forwardRef<
             {isGridReady && (
               <Grid size={12}>
                 {undoUI}
+                {redoUI}
                 <DataSheetGrid
                   ref={gridRef}
                   value={rowValues}
